renamed WrappedMiniMap to ResponsiveMiniMap
