+++ src/components/index.ts
@@ -46,6 +46,6 @@ export {
 } from './CollapsibleList/CollapsibleListEntry/CollapsibleListEntry'
 export { default as LinearProgressWithLabel } from './StyledLinearProgress/LinearProgressWithLabel'
 export { StyledLinearProgress } from './StyledLinearProgress/StyledLinearProgress'
-export { default as WrappedMiniMap } from './WrappedMiniMap/WrappedMiniMap'
+export { default as WrappedMiniMap } from './ResponsiveMiniMap/ResponsiveMiniMap'
 
 export * from './Nodes' // This stays as is because it's not following the Component/Component structure
